#91 Refactor container generation (move to service) and fix project from template
